+++ CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to the micro-web-service project will be documented in this 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[comment]: <> (## [Unreleased])
+## [Unreleased]
+### BREAKING CHANGES
+- The exported `Request` type is now a custom type using `Params`, `Unsure` from this package and `Query` from express.
+  This type is more type-safe than the `Request` exported by express.
+
+### Added
+- `ValidationRoute` for easy ajv validation setup.
+- `Unsure` type, an interface that sets all properties to `unknown`.
+- `Params` type, a more type-safe version of express' `ParamsDictionary`.
+
 ## [3.0.2] - 2021-04-06
 ### Fixed
 - Request logger losing class context while logging.